+++ venues/views.py
@@ -27,6 +27,6 @@ def venues():
 @blueprint.route("/venues/filters/<path:params>")
 def venues_filters(params):
     index_name = VENUES_INDEX
-    results = shared_filter_view(params, fields_dict, index_name)
+    results = shared_filter_view(request, params, fields_dict, index_name)
     filters_schema = FiltersWrapperSchema()
     return filters_schema.dump(results)