+++ works/views.py
@@ -38,6 +38,6 @@ def works():
 @blueprint.route("/works/filters/<path:params>")
 def works_filters(params):
     index_name = WORKS_INDEX
-    results = shared_filter_view(params, fields_dict, index_name)
+    results = shared_filter_view(request, params, fields_dict, index_name)
     filters_schema = FiltersWrapperSchema()
     return filters_schema.dump(results)